Add files via upload
Windows 11 is the latest major release of Microsoft's Windows NT operating system, released in October 2021. It is a free upgrade to its predecessor, Windows 10 (2015), available for any Windows 10 devices that meet the new Windows 11 system requirements.

Windows 11 features major changes to the Windows shell influenced by the canceled Windows 10X, including a redesigned Start menu, the replacement of its "live tiles" with a separate "Widgets" panel on the taskbar, the ability to create tiled sets of windows that can be minimized and restored from the taskbar as a group, and new gaming technologies inherited from Xbox Series X and Series S such as Auto HDR and DirectStorage on compatible hardware. Internet Explorer (IE) has been replaced by the Chromium-based Microsoft Edge as the default web browser like its predecessor, Windows 10, and Microsoft Teams is integrated into the Windows shell. Microsoft also announced plans to allow more flexibility in software that can be distributed via Microsoft Store, and to support Android apps on Windows 11 (including a partnership with Amazon to make its app store available for the function).

Citing security considerations, the system requirements for Windows 11 were increased over Windows 10. Microsoft only officially supports the operating system on devices using an eighth-generation Intel Core CPU or newer (with some minor exceptions), AMD Ryzen CPU based on Zen+ microarchitecture or newer, or a Qualcomm Snapdragon 850 ARM system-on-chip or newer, with UEFI secure boot and Trusted Platform Module (TPM) 2.0 supported and enabled (although Microsoft may provide exceptions to the TPM 2.0 requirement for OEMs). While the OS can be installed on unsupported processors, Microsoft does not guarantee the availability of updates. Windows 11 removed support for 32-bit x86 CPUs and devices which use BIOS firmware.

Windows 11 received a mixed reception at launch. Pre-release coverage of the operating system focused on its stricter hardware requirements, with discussions over whether they were legitimately intended to improve the security of Windows or as a ploy to upsell customers to newer devices, and over e-waste associated with the changes. Upon release, it was praised for its improved visual design, window management, and a stronger focus on security, but was criticized for various modifications to aspects of its user interface which were seen as worse than its predecessor.

As of October 2022, Windows 11, at 15%[15] (and up to 16%, in the US,[16] or more in some countries),[17] is the 2nd most popular Windows release on Windows PC,[18] while Windows 10 is still by far most popular with 5.5× the market share of its successor (and also 6.5× the share of Windows 7). Windows 11 is ahead of Windows 7, but after Windows 10 which stands at 72% globally, and 77% in the United States.[19] Across all platforms Windows 11 runs on a share of 4.0% of PCs.